ESTS-137690 Use q3stable dependencies where possible (#387)
All com.dell.cpsd dependencies in the q3table branches should point to
q3stable versions where they are available. If there is no q3stable
version available it means that it is an old version that is no longer
being built by master or q3stable branches and will not change in the
future. If there is a q3stable version it is being built by a q3stable
branch and we must consume that.